解决Socket异步传输问题事件

+++ ProducerServer/NIOClient/Server.cs
@@ -1,13 +1,11 @@
 ﻿using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Net;
 using System.Net.Sockets;
 using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
 
-namespace BrokerServer.NIOClient
+namespace ProducerServer.NIOClient
 {
     public class Server
     {
@@ -19,26 +17,23 @@ namespace BrokerServer.NIOClient
                                         // pool of reusable SocketAsyncEventArgs objects for write, read and accept socket operations
         SocketAsyncEventArgsPool m_receivePool;
         SocketAsyncEventArgsPool m_sendPool;
-        int m_totalBytesRead;           // counter of the total # bytes received by the server
         int m_numConnectedSockets;      // the total number of clients connected to the server
         Semaphore m_maxNumberAcceptedClients;
-        public event OnReceiveData ExecuteReceiveData;
+        private event OnReceiveData ExecuteReceiveData;
 
         public Server(int numConnections, int receiveBufferSize)
         {
-            m_totalBytesRead = 0;
             m_numConnectedSockets = 0;
             m_numConnections = numConnections;
             m_receiveBufferSize = receiveBufferSize;
-            
+
             m_bufferManager = new BufferManager(receiveBufferSize * numConnections * opsToPreAlloc,
                 receiveBufferSize);
             m_receivePool = new SocketAsyncEventArgsPool(numConnections);
             m_sendPool = new SocketAsyncEventArgsPool(numConnections);
             m_maxNumberAcceptedClients = new Semaphore(numConnections, numConnections);
         }
-
-        private void Init()
+        public void Init()
         {
             m_bufferManager.InitBuffer();
             for (int i = 0; i < m_numConnections; i++)
@@ -55,17 +50,22 @@ namespace BrokerServer.NIOClient
             }
         }
 
-        public void Start(IPEndPoint localEndPoint)
+        public void Start(IPEndPoint localEndPoint, IPEndPoint ipEnd = null)
         {
-            Init();
+            if(m_sendPool.Count == 0)
+            {
+                Init();
+            }
             listenSocket = new Socket(localEndPoint.AddressFamily, SocketType.Stream, ProtocolType.Tcp);
             listenSocket.Bind(localEndPoint);
+            listenSocket.Connect(ipEnd);
             listenSocket.Listen(100);
             StartAccept(null);
         }
         public void Connect(IPEndPoint ipEnd)
         {
-            listenSocket.Connect(ipEnd);
+            var result = listenSocket.BeginConnect(ipEnd, null, null);
+            listenSocket.EndConnect(result);
         }
         public void StartAccept(SocketAsyncEventArgs acceptEventArg)
         {
@@ -107,7 +107,6 @@ namespace BrokerServer.NIOClient
             {
                 ProcessReceive(readEventArgs);
             }
-
             StartAccept(e);
         }
         void IO_Completed(object sender, SocketAsyncEventArgs e)
@@ -139,23 +138,16 @@ namespace BrokerServer.NIOClient
                     //判断包的长度  
                     byte[] lenBytes = token.Buffer.GetRange(0, 4).ToArray();
                     int packageLen = BitConverter.ToInt32(lenBytes, 0);
+                    if (packageLen == 0) break;
                     if (packageLen > token.Buffer.Count - 4) break;
                     //包够长时,则提取出来,交给后面的程序去处理  
                     byte[] rev = token.Buffer.GetRange(4, packageLen).ToArray();
-                    String result = UnicodeEncoding.UTF8.GetString(rev);
+                    String result = UnicodeEncoding.Unicode.GetString(rev);
                     //从数据池中移除这组数据  
                     token.Buffer.RemoveRange(0, packageLen + 4);
                     //将数据包交给后台处理,这里你也可以新开个线程来处理.加快速度.  
                     if (ExecuteReceiveData != null)
-                    {
-                        Task.Run(() =>
-                        {
-                            return ExecuteReceiveData.Invoke(result);
-                        }).ContinueWith(result =>
-                        {
-                            ExecuteSendAsyncCallBack(token, result.Result);
-                        });
-                    }
+                        ExecuteReceiveData.BeginInvoke(token, result, ExecuteSendAsyncCallBack, token);
                     //这里API处理完后,并没有返回结果,当然结果是要返回的,却不是在这里, 这里的代码只管接收.  
                     //若要返回结果,可在API处理中调用此类对象的SendMessage方法,统一打包发送.不要被微软的示例给迷惑了.  
                 };
@@ -168,15 +160,15 @@ namespace BrokerServer.NIOClient
                 CloseClientSocket(e);
             }
         }
-        private void ExecuteSendAsyncCallBack(AsyncUserToken token, String sendData)
+        private void ExecuteSendAsyncCallBack(IAsyncResult result)
         {
-            //AsyncUserToken token = (AsyncUserToken)result.AsyncState;
-            //String sendData = ExecuteReceiveData.EndInvoke(result);
-            if (String.IsNullOrEmpty(sendData)) return;
+            AsyncUserToken token = (AsyncUserToken)result.AsyncState;
+            String sendData = ExecuteReceiveData.EndInvoke(result);
             byte[] body = Encoding.UTF8.GetBytes(sendData);
             byte[] len = BitConverter.GetBytes(body.Length);
             byte[] buffer = new byte[body.Length + 4];
             Array.Copy(len, buffer, 4);
+            Array.Copy(body, 0, buffer, 4, body.Length);
             int size = body.Length + 4, lenSize = 0;
             while (size > 0)
             {
@@ -238,22 +230,20 @@ namespace BrokerServer.NIOClient
     /// </summary>  
     /// <param name="token">客户端</param>  
     /// <param name="buff">客户端数据</param>
-    public delegate String OnReceiveData(String buff);
+    public delegate String OnReceiveData(AsyncUserToken token, String buff);
     public class SocketAsyncEventArgsPool
     {
         private Stack<SocketAsyncEventArgs> m_pool;
-        private Int32 m_count;
-        
-        public SocketAsyncEventArgsPool(int capacity)
+        public static SocketAsyncEventArgsPool InstanceSend = new SocketAsyncEventArgsPool();
+        public SocketAsyncEventArgsPool(int maxNum = 10)
         {
-            m_pool = new Stack<SocketAsyncEventArgs>(capacity);
-            m_count = 0;
+            m_pool = new Stack<SocketAsyncEventArgs>(maxNum);
         }
         private Int32 m_pushLock = 0;
         public void Push(SocketAsyncEventArgs item)
         {
             if (item == null) { throw new ArgumentNullException("Items added to a SocketAsyncEventArgsPool cannot be null"); }
-            while(Interlocked.CompareExchange(ref m_pushLock, 1, 0) == 0)
+            while (Interlocked.CompareExchange(ref m_pushLock, 1, 0) == 0)
             {
                 Thread.Sleep(1);
             }
@@ -274,7 +264,7 @@ namespace BrokerServer.NIOClient
                 {
                     Thread.Sleep(1);
                 }
-                if (m_pool.Count != 0) 
+                if (m_pool.Count != 0)
                 {
                     item = m_pool.Pop();
                 }
@@ -282,6 +272,37 @@ namespace BrokerServer.NIOClient
                 if (item != null) break;
                 Thread.Sleep(1);
             }
+            Volatile.Write(ref m_popLock, 0);
+            return item;
+        }
+        public SocketAsyncEventArgs PopOrNew(Func<SocketAsyncEventArgs> func)
+        {
+            while (Interlocked.CompareExchange(ref m_popLock, 1, 0) == 0)
+            {
+                Thread.Sleep(1);
+            }
+            SocketAsyncEventArgs item = null;
+            while (Interlocked.CompareExchange(ref m_pushLock, 1, 0) == 0)
+            {
+                Thread.Sleep(1);
+            }
+            if (m_pool.Count != 0)
+            {
+                item = m_pool.Pop();
+            }
+            else
+            {
+                if(func == null)
+                {
+                    item = new SocketAsyncEventArgs();
+                }
+                else
+                {
+                    item = func();
+                }
+                m_pool.Push(item);
+            }
+            Volatile.Write(ref m_pushLock, 0);
             Volatile.Write(ref m_popLock, 0);
             return item;
         }
